Add files via upload
This project is on POSTGRE SQL, for Music Data Analysis. 
A comprehensive analysis was conducted through a series of queries to unveil insights from the database. By leveraging PostgreSQL's robust querying capabilities, various aspects of the dataset were explored, including but not limited to data distribution, trends, and relationships. Through aggregation functions, joins, and subqueries, key metrics such as averages, counts, and sums were calculated to provide a deeper understanding of the underlying data patterns. Additionally, advanced SQL techniques were employed to extract meaningful information, such as identifying outliers, detecting correlations, and segmenting data based on specific criteria. This analysis not only sheds light on the current state of the dataset but also lays the foundation for informed decision-making and future exploration of the data.